requirements.txt
altair==5.5.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560565029/work
charset-normalizer==3.4.2
click==8.1.8
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1731428353116/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1737269765476/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
et_xmlfile==2.0.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
filelock==3.18.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1738940329223/work
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1738421264/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1734823942230/work
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1736908211538/work
lxml @ file:///Users/runner/miniforge3/conda-bld/lxml_1739211586271/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219691421/work
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mpmath==1.3.0
munkres==1.1.4
narwhals==1.38.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1737331087001/work/dist/numpy-2.2.2-cp312-cp312-macosx_10_13_x86_64.whl#sha256=70b38e39528310cd817dabbc387ffca2ce641bfe0fe8d86e396a51dd03f1d376
openpyxl==3.1.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1726878436460/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1735929764289/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==6.30.2
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1735327393051/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyaml==25.1.0
pyarrow==20.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
pyreadr @ file:///Users/runner/miniforge3/conda-bld/pyreadr_1726940170180/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1738271043357/work
RapidFuzz==3.12.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rpy2 @ file:///Users/runner/miniforge3/conda-bld/rpy2_1696426071713/work
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1736617276787/work/dist/scipy-1.15.1-cp312-cp312-macosx_10_13_x86_64.whl#sha256=6266da2cf22a0e9d6e56239a4af131c9761876c475dc7150920d07965e4dd298
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
setuptools==75.8.0
simplegeneric @ file:///home/conda/feedstock_root/build_artifacts/simplegeneric_1734685420077/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sklearn-compat==0.1.3
smmap==5.0.2
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1727986709373/work
streamlit==1.45.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1732615945596/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
tzlocal @ file:///Users/runner/miniforge3/conda-bld/tzlocal_1724952050955/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1736692573104/work
urllib3==2.4.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
wheel==0.45.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
